+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- `/fi rename "new name"` command allowing fief owners to rename their fief (`fiefs.rename`)
+
 ## [0.11.0]
 
 ### Added